xcorr: cross correlation analysis
From a given dataframe (of telemetry) the XCorr transformer
tests the predictive power of each variable on one another.
Each column is predicted thanks to all the others, one by one.
The feature importance map is thus built and a heatmap of
dependencies can be draw.

The dependency is with respect to feature importances output by XGBoost Regressor.

TODO: retain the confidence factor (test set error)