viz: two new features - Head-Up Display and Search
The HUD is displayed as a HTML5 element on top of the 3D graph canva.

The Search is a simple field on top left with a datalist for
auto-complete. It will display strings as:
            <NODE.ID>:<NODE.NAME>/<NODE.GROUP>

It listens to "Enter" press event to look for the node.id present in the
input.

WIP for the ctrl+ENTER press event to highlight all nodes which contains
search-string.