fetch: using dates from pandas only - fix #59
Avoiding to translate pandas timestamps to datetime.datetime.
This solves pylint recursion and
Closes #59

More tests where added to test the build start and end date function for fetching
in case it would be called from code, so with different kinds of objects than None and Strings.
So actual datetime.datetime or pandas timestamps.