Pin astroid version to fix pylint recursion error
This fixes a problem we've encountered where pylint tests fail with
this error:

	RecursionError: maximum recursion depth exceeded in comparison

The bug appears to be this one:

	https://github.com/PyCQA/pylint/issues/3318

For the record, this was extremely hard to track down.  The error
appeared erratically, as there appears to be some caching (and
invalidation + recreation) of tox environments and/or packages that
feeds into this.

We can revisit the version pinning in the future; at the moment,
there's no other motivation for the pinning.

Signed-off-by: Hugh Brown (Saint Aardvark the Carpeted) <aardvark@saintaardvarkthecarpeted.com>