Set default graph_link_threshold in click options
It turns out that if an argument is:

a) specified by click, but has no default; and
b) included in the function signature with a reasonable default; and
c) you invoke the command and don't provide the argument;

then the click library sets the argument to `None`, and the default
argument in the function signature is ignored.  This fixes that problem.

Signed-off-by: Hugh Brown (Saint Aardvark the Carpeted) <aardvark@saintaardvarkthecarpeted.com>