Add PolarisGraph class + tests, use it in writing heatmap
This adds PolarisGraph, a new class for Polaris dependency graphs.
This is a refactoring of the code in writers.py, but also adds
metadata (at last!) to the output.  It also adds unit tests for the
class, and uses the class when writing the dependency graph at the end
of `polaris learn`.

Signed-off-by: Hugh Brown (Saint Aardvark the Carpeted) <aardvark@saintaardvarkthecarpeted.com>