Add loadGraphData() to the UI to support old + new graph files
Previously, we had constructed the Graph object by passing the graph
file's URL to ForceGraph3D's `loadData()` function; that would
retrieve the data automatically.  However, that data needs to be in
the format:

```
{ "nodes": [ ... ],
  "links": [ ... ]
}
```

Previously, the graph file produced by `polaris learn` was in exactly
this format, so passing the URL of the graph file to `loadData()`
worked just fine.  However, the new PolarisGraph class saves data in
this format:

```
{ "metadata": { .. },
  "graph": { "nodes": [ ... ],
             "links": [ ... ],
           }
}
```

Thus, we need to massage the data a bit, and to handle both cases so
we can ensure backward compatibility.

loadGraphData() does that massaging.  It fetches a Polaris graph file
and returns the data suitable for construction of the 3D graph.  If
the fetched data has a "graph" key (as the new PolarisGraph class
produces), that will be returned; if that key is not present, it's
assumed the JSON is in the previous format, and the data is returned
as-is.

Finally, we change the construction of the Graph object by using
ForceGraph3D's `graphData()` function.

Signed-off-by: Hugh Brown (Saint Aardvark the Carpeted) <aardvark@saintaardvarkthecarpeted.com>