Add integration test for polaris fetch
All the data for the time range are stored in tests/test_data
This data is then transferred to a temporary folder where fetch
is tested. Since there is no data to fetch, there is no load on any
server.

This increases the coverage of polaris by a whopping 10% :flex:

Signed-off-by: Adithya Venkateswaran <adithyav1511@gmail.com>