Handle case in `polaris learn` where metadata is empty
When running `polaris learn` on CSV data, the metadata created when
loading the data won't have the `feature_columns` key.  Here, we
detect that, log it, and set `feature_columns` parameter for Xcorr to
an empty array.

This also adds a unit test to catch the case where the metadata
supplied as an argument to the XCorr function is empty.

Fixes #174

Signed-off-by: Hugh Brown (Saint Aardvark the Carpeted) <aardvark@saintaardvarkthecarpeted.com>