Issue #184: Temporarily disable two pylint warnings in extraction.py
As detailed in
https://gitlab.com/librespacefoundation/polaris/polaris/-/issues/184,
astroid 2.5.8 currently has a bug that causes it to issue false
warnings in Pylint, causing our pipeline to fail.  This only happened
with the 2.5.8 release -- the 2.5.7 version worked just fine before,
and if you reinstall it by force in a `tox` virtualenv, it works fine
again.

This is a temporary fix.  There are two larger issues to be addressed:

- First off, tox.ini tries to pin astroid to version 2.3.3 -- it's
unclear to me why this is being upgraded to 2.5.8.

- Second, of course, is the upstream bug.

This change will be reverted when one or both of these are fixed.  In
the meantime, this will get our pipeline passing again -- or, more
accurately, *not* throwing false errors and failing.

Signed-off-by: Hugh Brown (Saint Aardvark the Carpeted) <aardvark@saintaardvarkthecarpeted.com>