Refactor __repr__/__str__ methods for PolarisGraph and PolarisDataset
Among other things, the PolarisGraph and PolarisDataset classes are
meant to allow us to read and save files with graphs and datasets.
JSON is the format we've picked, so having the `__repr__()` and `__str__()`
methods return JSON strings makes sense.

The original code, though, was *cough* less than ideal: `__repr__()`
methods should return a string, after all, which was something I
hadn't realized before.  Recently, an update to astroid and Pylint
caught this shortcoming.

To resolve this, I have:

- put all the JSON-producing logic in the `to_json()` function for each
class

- then called `to_json()` from `__repr__()` and `__str__()`

Signed-off-by: Hugh Brown (Saint Aardvark the Carpeted) <aardvark@saintaardvarkthecarpeted.com>